Move toHaveSuspenseCacheEntryUsing matchers to global matchers
