+++ src/testing/matchers/index.ts
@@ -1,6 +1,8 @@
 import { expect } from '@jest/globals';
 import { toMatchDocument } from './toMatchDocument';
+import { toHaveSuspenseCacheEntryUsing } from './toHaveSuspenseCacheEntryUsing';
 
 expect.extend({
+  toHaveSuspenseCacheEntryUsing,
   toMatchDocument,
 });
